mcp/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "google-genai>=1.14.0",
    "mcp>=1.8.0",
    "python-dotenv>=1.1.0",
]

[tool.uv.workspace]
members = [
    "mcp-client",
]
